linux_local_container: Add linux_local_container
The difference between a LinuxContainer and a LinuxLocalContainer
is that the latter installs packages in a separate subdirectory
but does not run commands through proot. Instead, it sets up
the library and runtime paths and runs the command with those
environment variables. This is useful if the overhead of using
proot is not required.

bumpversion: minor